refactor: split one file into multiple different business logic files
