+++ pkg/deviceplugins/usb_device_plugin.go
@@ -20,7 +20,6 @@ package deviceplugins
  */
 
 import (
-	"bufio"
 	"context"
 	"errors"
 	"fmt"
@@ -28,7 +27,6 @@ import (
 	"os"
 	"path"
 	"path/filepath"
-	"strconv"
 	"strings"
 	"sync"
 	"time"
@@ -449,45 +447,6 @@ func (plugin *USBDevicePlugin) PreStartContainer(context.Context, *pluginapi.Pre
 	return &pluginapi.PreStartContainerResponse{}, nil
 }
 
-type LocalDevices struct {
-	// For quicker indexing, map devices based on vendor string
-	devices map[int][]*USBDevice
-}
-
-// finds by vendor and product
-func (l *LocalDevices) find(vendor, product int) *USBDevice {
-	if devices, exist := l.devices[vendor]; exist {
-		for _, local := range devices {
-			if local.Product == product {
-				return local
-			}
-		}
-	}
-	return nil
-}
-
-// remove all cached elements
-func (l *LocalDevices) remove(usbdevs []*USBDevice) {
-	for _, dev := range usbdevs {
-		devices, exists := l.devices[dev.Vendor]
-		if !exists {
-			continue
-		}
-
-		for i, usb := range devices {
-			if usb.GetID() == dev.GetID() {
-				devices = append(devices[:i], devices[i+1:]...)
-				break
-			}
-		}
-
-		l.devices[dev.Vendor] = devices
-		if len(devices) == 0 {
-			delete(l.devices, dev.Vendor)
-		}
-	}
-}
-
 // return a list of USBDevices while removing it from the list of local devices
 func (l *LocalDevices) fetch(selectors []v1.USBSelector) ([]*USBDevice, bool) {
 	usbdevs := make([]*USBDevice, len(selectors))
@@ -514,77 +473,6 @@ func (l *LocalDevices) fetch(selectors []v1.USBSelector) ([]*USBDevice, bool) {
 	return usbdevs, true
 }
 
-func discoverPluggedUSBDevices() *LocalDevices {
-	usbDevices := make(map[int][]*USBDevice, 0)
-	err := filepath.Walk(pathToUSBDevices, func(path string, info os.FileInfo, err error) error {
-		if err != nil {
-			return err
-		}
-		// Ignore named usb controllers
-		if strings.HasPrefix(info.Name(), "usb") {
-			return nil
-		}
-		// We are interested in actual USB devices information that
-		// contains idVendor and idProduct. We can skip all others.
-		if _, err := os.Stat(filepath.Join(path, "idVendor")); err != nil {
-			return nil
-		}
-
-		// Get device information
-		if device := parseSysUeventFile(path); device != nil {
-			usbDevices[device.Vendor] = append(usbDevices[device.Vendor], device)
-		}
-		return nil
-	})
-
-	if err != nil {
-		log.Log.Reason(err).Error("Failed when walking usb devices tree")
-	}
-	return &LocalDevices{devices: usbDevices}
-}
-
-func parseSelector(s *v1.USBSelector) (int, int, error) {
-	val, err := strconv.ParseInt(s.Vendor, 16, 32)
-	if err != nil {
-		return -1, -1, err
-	}
-	vendor := int(val)
-
-	val, err = strconv.ParseInt(s.Product, 16, 32)
-	if err != nil {
-		return -1, -1, err
-	}
-	product := int(val)
-
-	return vendor, product, nil
-}
-
-func DiscoverAllowedUSBDevices(usbs []v1.USBHostDevice) map[string][]*PluginDevices {
-	// The return value: USB USBDevice Plugins found and permitted to be exposed
-	plugins := make(map[string][]*PluginDevices)
-	// All USB devices found plugged in the Node
-	localDevices := discoverLocalUSBDevicesFunc()
-	for _, usbConfig := range usbs {
-		resourceName := usbConfig.ResourceName
-		// only accept ExternalResourceProvider: true for USB devices
-		if !usbConfig.ExternalResourceProvider {
-			log.Log.V(6).Infof("Skipping discovery of %s. To be handled by external device-plugin",
-				resourceName)
-			continue
-		}
-		index := 0
-		usbdevs, foundAll := localDevices.fetch(usbConfig.Selectors)
-		for foundAll {
-			// Create new USB USBDevice Plugin with found USB Devices for this resource name
-			pluginDevices := newPluginDevices(resourceName, index, usbdevs)
-			plugins[resourceName] = append(plugins[resourceName], pluginDevices)
-			index++
-			usbdevs, foundAll = localDevices.fetch(usbConfig.Selectors)
-		}
-	}
-	return plugins
-}
-
 func NewUSBDevicePlugin(resourceName string, pluginDevices []*PluginDevices) *USBDevicePlugin {
 	s := strings.Split(resourceName, "/")
 	resourceID := s[0]
@@ -600,131 +488,4 @@ func NewUSBDevicePlugin(resourceName string, pluginDevices []*PluginDevices) *US
 		initialized:  false,
 		lock:         &sync.Mutex{},
 	}
-}
-
-func parseSysUeventFile(path string) *USBDevice {
-	link, err := os.Readlink(path)
-	if err != nil {
-		return nil
-	}
-
-	// Grab all details we are interested from uevent
-	file, err := os.Open(filepath.Join(path, "uevent"))
-	if err != nil {
-		logrus.Printf("Unable to access %s/%s\n", path, "uevent")
-		return nil
-	}
-	defer file.Close()
-
-	u := USBDevice{
-		PCIAddress: parseUSBSymLinkToPCIAddress(link),
-	}
-
-	scanner := bufio.NewScanner(file)
-	for scanner.Scan() {
-		line := scanner.Text()
-		values := strings.Split(line, "=")
-		if len(values) != 2 {
-			logrus.Printf("Skipping %s due not being key=value\n", line)
-			continue
-		}
-
-		key, value := values[0], values[1]
-		if !parseSysUeventKeyValue(key, value, &u) {
-			return nil
-		}
-	}
-
-	return &u
-}
-
-func parseSysUeventKeyValue(key string, value string, u *USBDevice) bool {
-	switch key {
-	case "BUSNUM":
-		val, err := strconv.ParseInt(value, 10, 32)
-		if err != nil {
-			return false
-		}
-		u.Bus = int(val)
-	case "DEVNUM":
-		val, err := strconv.ParseInt(value, 10, 32)
-		if err != nil {
-			return false
-		}
-		u.DeviceNumber = int(val)
-	case "PRODUCT":
-		products := strings.Split(value, "/")
-		if len(products) != 3 {
-			return false
-		}
-
-		val, err := strconv.ParseInt(products[0], 16, 32)
-		if err != nil {
-			return false
-		}
-		u.Vendor = int(val)
-
-		val, err = strconv.ParseInt(products[1], 16, 32)
-		if err != nil {
-			return false
-		}
-		u.Product = int(val)
-
-		val, err = strconv.ParseInt(products[2], 16, 32)
-		if err != nil {
-			return false
-		}
-		u.BCD = int(val)
-	case "DEVNAME":
-		u.DevicePath = filepath.Join("/dev", value)
-	default:
-		logrus.Printf("Skipping unknown key=value %s=%s\n", key, value)
-	}
-
-	return true
-}
-
-func WalkUSBDevices() (map[int][]*USBDevice, error) {
-	usbDevices := make(map[int][]*USBDevice, 0)
-	err := filepath.Walk("/sys/bus/usb/devices", func(path string, info os.FileInfo, err error) error {
-		if err != nil {
-			return err
-		}
-
-		// Ignore named usb controllers
-		if strings.HasPrefix(info.Name(), "usb") {
-			return nil
-		}
-		// We are interested in actual USB devices information that
-		// contains idVendor and idProduct. We can skip all others.
-		if _, err := os.Stat(filepath.Join(path, "idVendor")); err != nil {
-			return nil
-		}
-
-		//fmt.Println(path)
-		//fmt.Printf("%#v\n", info)
-
-		// Get device information
-		if device := parseSysUeventFile(path); device != nil {
-			usbDevices[device.Vendor] = append(usbDevices[device.Vendor], device)
-		}
-		return nil
-	})
-
-	if err != nil {
-		return nil, err
-	}
-
-	return usbDevices, nil
-}
-
-func parseUSBSymLinkToPCIAddress(link string) string {
-	paths := strings.Split(link, "/")
-
-	if len(paths) < 3 {
-		return ""
-	}
-
-	pciAddress := paths[len(paths)-3 : len(paths)-2][0]
-	return pciAddress
 }
