+++ pkg/deviceplugins/usb_device_plugin_helper_test.go
